requirements.txt
pandas
numpy
plotly
scikit-learn
pyyaml
uploaded_file